requirements.txt
asgiref==3.5.2
Babel==2.10.1
boto3==1.23.4
botocore==1.26.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
django-money==2.1.1
django-on-heroku==1.1.2
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docker==5.0.3
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
Pillow==9.1.1
psycopg2-binary==2.9.3
py-moneyed==1.2
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
websocket-client==1.3.2
whitenoise==6.1.0
